feat: Add schema.Invalidate
