+++ .changeset/nine-spies-repair.md
@@ -1,0 +1,5 @@
+---
+'@rest-hooks/normalizr': minor
+---
+
+Support schemas without denormalizeOnly and no denormalize method
