+++ .changeset/wise-files-confess.md
@@ -1,0 +1,7 @@
+---
+'@rest-hooks/endpoint': minor
+'@rest-hooks/graphql': minor
+'@rest-hooks/rest': minor
+---
+
+Add schema.Invalidate (new name for Delete)
