+++ examples/benchmark/README.md
@@ -26,3 +26,8 @@ Performance compared to normalizr package (higher is better):
 | normalize (long)    | 72%      | 72%        |
 | denormalize (long)  | 100%     | 830%       |
 | denormalize (short) | 600%     | 1120%      |
+
+[Comparison done on a Ryzen 7950x; Ubuntu; Node 18.15.0]
+
+Our normalize is slower due to handling much more features like metadata. Denormalize
+is even more feature rich, but significantly faster.
