+++ packages/endpoint/src/schema.js
@@ -5,4 +5,5 @@ export { default as Array } from './schemas/Array.js';
 export { default as All } from './schemas/All.js';
 export { default as Object } from './schemas/Object.js';
 export { default as Delete } from './schemas/Delete.js';
+export { default as Invalidate } from './schemas/Invalidate.js';
 export { default as Entity } from './schemas/EntitySchema.js';
