+++ packages/endpoint/src/special.ts
@@ -1,1 +1,2 @@
 export const DELETED = Symbol('ENTITY WAS DELETED');
+export const INVALID = DELETED;
